travel-ranker/requirements.txt
streamlit==1.31.0
pandas==2.2.0
numpy==1.26.0
requests==2.31.0
python-dotenv==1.0.0
amadeus==8.1.0
humanize==4.9.0
pytest==8.0.0
streamlit-aggrid==0.3.4
streamlit-extras==0.3.6
